Fix Dependency Scopes (#30)
The `PipelineStepCollection` was being registered as a singleton, but
because it injects `IServiceProvider` for the `IStepProvider` interface,
it can't be used to resolve scoped dependencies like steps (the
`IServiceProvider` that gets injected is the root singleton version, not
the scoped version).

We can't make the `PipelineStepCollection` a scoped service becasuse the
`IPipelineStepCollector` interface is used by
`PipelineApplication.UseStep<T>`, which runs outside of any given scope
(scopes are for each pipeline _execution_, not for setup).

Splitting out the provider and the collector into two separate classes
lets us keep the provider a scoped service, while allowing the default
collector to be a singleton. It actually looks a bit cleaner like this
anyway.