Add new file system functionality (#37)
While developing another build, I need to load an `IDirectory` outside
of the current directory (the staging directory provided by the build
agent) and realised there was no way to do that.

I also needed to get the name of something based on a file name, minus
its extension.

Neither of those features were supported, so... here they are.

---------

Co-authored-by: Duncan Blair <867976+nimro@users.noreply.github.com>