Refactor Pipeline Runner (#40)
Refactors the pipeline runner to be more flexible:
- Each step is run inside a self-contained `RunStep` function making it
easier to reason about.
- `RunSteps` is responsible for wrangling all the steps and deciding
when to break/continue.
- `RunPipeline` is responsible for running the steps and compiling a
summary.
- `Computing the `ExitCode` is now done in one place based on all the
summary information, making it easier to test in isolation, and easier
to reason about as it doesn't change throughout the pipeline execution.

The big breaking change with this approach is that when stopping on
error, the exception isn't thrown, because then that breaks all of the
step summary info. It's not great if we lose all our output because of
an unhandled exception. Considering the context in which Hamelin will be
used, I think it's okay for us to just log the exception and ensure the
exit code is set correctly.