RunWithExitCode And Other QoL (#53)
Setting `Environment.ExitCode` doesn't play nicely in anything outside
of a standalone application. For example, `System.CommandLine` expects
you to return your exit code from the handler delegate.

My goal in this PR was to have one function that runs the pipeline and
returns an `int` based on whether it was successful or not. The `Run`
and `RunAsync` methods are already extension methods that wrap some of
this, so I decided to copy those, store the pipeline result after the
post pipeline hooks, then retrieve it after the fact.

My approach feels a bit gross - especially having to register the
summary store as a singleton instead of scoped, but I couldn't think of
a better way to do it, so feedback would be greatly appreciated.

I also wanted to solve another problem: adding hooks manually is a bit
of a pain, and not very discoverable. Having specific extension methods
makes them a little bit easier, as well as having one method just to
scoop up all the hooks in an assembly.