Update to .NET 10 (#54)
* Projects now target 10.*
* SDK references upgraded to to 10.*
* Extension methods now use new `extension` block syntax.
* Fixed a bug in the `PipelineHostTests` - We technically had a race
condition because `BackgroundService.StartAsync` just starts the task,
it doesn't wait for it to complete. The behaviour of this method changed
in .NET 10, so sometimes one of the tests would fail when run in a
group. We're now correctly waiting for the internal task to complete
before asserting the output.